smartcare/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
paho-mqtt==1.6.1
Pillow==10.1.0
tensorflow==2.20.0
numpy==1.24.3
requests==2.31.0
python-decouple==3.8
channels==4.0.0
daphne==4.0.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
scikit-learn
matplotlib
pandas
keras==2.20.0
